# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vending Machine Orchestrator API",
        "description": "RESTful API for managing vending machines and a shared product inventory. Machines operate as state machines coordinated by an orchestrator.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development"
        }
    ],
    "paths": {
...
        "/api/vending-machines": {
            "get": {
                "tags": [
                    "Vending Machines"
                ],
                "summary": "List all vending machines",
                "operationId": "d9c0a55bbca6796c336d85024f306718",
                "responses": {
                    "200": {
                        "description": "List of vending machines",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VendingMachine"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vending Machines"
                ],
                "summary": "Create a new vending machine",
                "operationId": "b68057e9a0efb1fe50fd02294c03035b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Machine A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Machine created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VendingMachine"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/vending-machines/{id}": {
            "get": {
                "tags": [
                    "Vending Machines"
                ],
                "summary": "Get a vending machine by ID",
                "operationId": "ec339afdbfbf55698f430e7c2ca47293",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vending machine details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VendingMachine"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Machine not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vending Machines"
                ],
                "summary": "Update a vending machine",
                "operationId": "e6888f84691489ee07aa74ede6bd89fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Machine A (Updated)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Machine updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VendingMachine"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Machine not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vending Machines"
                ],
                "summary": "Delete a vending machine",
                "operationId": "d2f85355269775949f69835f00ffd22f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Machine deleted"
                    },
                    "404": {
                        "description": "Machine not found"
                    }
                }
            }
        },
        "/api/vending-machines/{id}/reset": {
            "post": {
                "tags": [
                    "Vending Machines"
                ],
                "summary": "Reset a vending machine to idle state",
                "description": "Forces a vending machine back to idle state regardless of its current state. Useful for recovering stuck machines.",
                "operationId": "c978a9be2cc5afc3c94fc2521ef93ab6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine reset to idle",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Machine has been reset to idle state."
                                        },
                                        "machine": {
                                            "$ref": "#/components/schemas/VendingMachine"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Machine not found"
                    },
                    "409": {
                        "description": "Machine is already idle",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Machine is already idle."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "VendingMachineStatus": {
                "description": "The operational state of a vending machine.",
                "type": "string",
                "enum": [
                    "idle",
                    "choose_product",
                    "processing"
                ]
            },
...
            "VendingMachine": {
                "required": [
                    "id",
                    "name",
                    "status",
                    "usage_count",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Machine A"
                    },
                    "status": {
                        "$ref": "#/components/schemas/VendingMachineStatus"
                    },
                    "usage_count": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}